[Add] refactoring for the LMCache MP cache engine
