+++ lmcache/v1/multiprocess/http_server.py
@@ -61,12 +61,9 @@ async def lifespan(app: FastAPI):
         torch_dev.is_available(),
     )
     mp_config = _configs["mp"]
-    if mp_config.engine_type == "blend":
-        # First Party
-        from lmcache.v1.multiprocess.blend_server_v2 import run_cache_server
-    else:
-        # First Party
-        from lmcache.v1.multiprocess.server import run_cache_server
+
+    # First Party
+    from lmcache.v1.multiprocess.server import run_cache_server
 
     result = run_cache_server(
         mp_config=mp_config,
