+++ lmcache/v1/multiprocess/modules/gpu_transfer.py
@@ -1,0 +1,640 @@
+# SPDX-License-Identifier: Apache-2.0
+"""GPU-based KV cache transfer operations for the MPCacheEngine."""
+
+# Standard
+from dataclasses import dataclass
+from itertools import islice
+from typing import Generator
+import time
+
+# First Party
+from lmcache import torch_dev, torch_device_type
+from lmcache.logging import init_logger
+from lmcache.utils import (
+    EngineType,
+    _lmcache_nvtx_annotate,
+    check_interprocess_event_support,
+)
+from lmcache.v1.distributed.api import (
+    MemoryLayoutDesc,
+    ObjectKey,
+)
+from lmcache.v1.gpu_connector.gpu_ops import (
+    lmcache_memcpy_async_d2h,
+    lmcache_memcpy_async_h2d,
+)
+from lmcache.v1.gpu_connector.utils import LayoutHints
+from lmcache.v1.memory_management import MemoryObj
+from lmcache.v1.mp_observability.event import Event, EventType
+from lmcache.v1.multiprocess.custom_types import (
+    IPCCacheEngineKey,
+    KVCache,
+)
+from lmcache.v1.multiprocess.engine_context import MPCacheEngineContext
+from lmcache.v1.multiprocess.engine_module import (
+    HandlerSpec,
+    ThreadPoolType,
+)
+from lmcache.v1.multiprocess.gpu_context import GPUCacheContext
+from lmcache.v1.multiprocess.protocols.base import RequestType
+import lmcache.c_ops as lmc_ops
+
+logger = init_logger(__name__)
+
+
+def get_layout_desc(gpu_context: GPUCacheContext, num_tokens: int) -> MemoryLayoutDesc:
+    """Get the memory layout description for a given GPU context and number of tokens.
+
+    Supports multiple KV layer groups with different shapes and dtypes.
+
+    Args:
+        gpu_context: The GPU cache context containing the KV cache information.
+        num_tokens: The number of tokens to determine the layout for.
+
+    Returns:
+        MemoryLayoutDesc: The memory layout description containing shapes and dtypes.
+    """
+    num_groups = gpu_context.kv_layer_groups_manager.num_groups
+    shapes = [
+        gpu_context.get_kv_buffer_shape(num_tokens, group_idx)
+        for group_idx in range(num_groups)
+    ]
+    dtypes = [
+        gpu_context.kv_layer_groups_manager.kv_layer_groups[group_idx].dtype
+        for group_idx in range(num_groups)
+    ]
+    return MemoryLayoutDesc(shapes=shapes, dtypes=dtypes)
+
+
+def batched_iteration(lst: list, batch_size: int) -> Generator[tuple, None, None]:
+    """Utility function to iterate over a list in batches.
+
+    Args:
+        lst: The list to iterate over.
+        batch_size: The size of each batch.
+
+    Yields:
+        Batches of the list as tuples.
+
+    Raises:
+        ValueError: If batch_size is less than 1.
+    """
+    if batch_size < 1:
+        raise ValueError("batch size must be at least one")
+    it = iter(lst)
+    while batch := tuple(islice(it, batch_size)):
+        yield batch
+
+
+@dataclass
+class GPUContextEntry:
+    """Registered GPU context metadata for a single worker instance.
+
+    Args:
+        gpu_context: The GPU cache context managing shape and pointers
+            to vLLM GPU KV cache tensors.
+        model_name: The name of the model associated with this KV cache.
+        world_size: The world size associated with this KV cache.
+    """
+
+    gpu_context: GPUCacheContext
+    model_name: str
+    world_size: int
+
+
+class GPUTransferModule:
+    """Handles GPU-based KV cache transfer operations.
+
+    Owns GPU context registrations and provides handlers for
+    register, unregister, store, and retrieve of GPU KV caches.
+
+    Args:
+        ctx: The shared engine context.
+    """
+
+    def __init__(self, ctx: MPCacheEngineContext) -> None:
+        self._ctx = ctx
+        self._gpu_contexts: dict[int, GPUContextEntry] = {}
+
+    def get_handlers(self) -> list[HandlerSpec]:
+        """Return handler specs for all request types this module serves.
+
+        Returns:
+            A list of HandlerSpec entries mapping request types to
+            their handler callables and thread pool assignments.
+        """
+        return [
+            HandlerSpec(
+                RequestType.REGISTER_KV_CACHE,
+                self.register_kv_cache,
+                ThreadPoolType.SYNC,
+            ),
+            HandlerSpec(
+                RequestType.UNREGISTER_KV_CACHE,
+                self.unregister_kv_cache,
+                ThreadPoolType.SYNC,
+            ),
+            HandlerSpec(
+                RequestType.STORE,
+                self.store,
+                ThreadPoolType.AFFINITY,
+            ),
+            HandlerSpec(
+                RequestType.RETRIEVE,
+                self.retrieve,
+                ThreadPoolType.AFFINITY,
+            ),
+        ]
+
+    def report_status(self) -> dict:
+        """Return GPU transfer module status information.
+
+        Returns:
+            A dict containing registered GPU instance IDs and
+            per-instance KV cache layout metadata.
+        """
+        registered_gpu_ids: list[int] = []
+        gpu_context_meta: dict[str, dict] = {}
+
+        for instance_id, entry in self._gpu_contexts.items():
+            registered_gpu_ids.append(instance_id)
+            ctx = entry.gpu_context
+            gpu_context_meta[str(instance_id)] = {
+                "model_name": entry.model_name,
+                "world_size": entry.world_size,
+                "kv_cache_layout": {
+                    "num_layers": ctx.num_layers,
+                    "inference_engine_logical_block_size": (
+                        ctx.kv_layer_groups_manager.inference_engine_logical_block_size
+                    ),
+                    "group_physical_block_sizes": ctx.group_physical_block_sizes,
+                    "group_compress_ratios": ctx.group_compress_ratios,
+                    "hidden_dim_sizes": str(ctx.hidden_dim_sizes),
+                    "dtype": str(ctx.dtype),
+                    "is_mla": ctx.is_mla,
+                    "num_blocks": ctx.num_blocks,
+                    "gpu_kv_format": ctx.gpu_kv_format_name,
+                    "gpu_kv_shape": ctx.gpu_kv_shape,
+                    "gpu_kv_concrete_shape": ctx.concrete_gpu_kv_shape,
+                    "attention_backend": ctx.attention_backend,
+                    "cache_size_per_token": ctx.cache_size_per_token(),
+                },
+            }
+
+        return {
+            "registered_gpu_ids": registered_gpu_ids,
+            "gpu_context_meta": gpu_context_meta,
+        }
+
+    def close(self) -> None:
+        """Release GPU resources owned by this module."""
+        had_contexts = len(self._gpu_contexts) > 0
+        self._gpu_contexts.clear()
+        if had_contexts:
+            torch_dev.empty_cache()
+
+    def register_kv_cache(
+        self,
+        instance_id: int,
+        kv_caches: KVCache,
+        model_name: str,
+        world_size: int,
+        engine_type: EngineType,
+        layout_hints: LayoutHints,
+    ) -> None:
+        """Register the KV cache tensors for a given GPU instance ID.
+
+        Args:
+            instance_id: The GPU instance ID (such as PID).
+            kv_caches: The KV cache tensor wrappers from the
+                serving engine.
+            model_name: The name of the model associated with this KV cache.
+            world_size: The world size associated with this KV cache.
+            engine_type: Which serving engine produced the caches.
+                Forwarded to GPUCacheContext for format detection.
+            layout_hints: See LayoutHints.  Forwarded to
+                GPUCacheContext for GPU KV format detection.
+        """
+        if instance_id in self._gpu_contexts:
+            logger.warning(
+                "Instance %s's KV cache is already registered, "
+                "skipping the new registration",
+                instance_id,
+            )
+            return
+
+        gpu_context = GPUCacheContext(
+            kv_caches,
+            self._ctx.chunk_size,
+            layout_hints=layout_hints or None,
+            engine_type=engine_type,
+        )
+        self._gpu_contexts[instance_id] = GPUContextEntry(
+            gpu_context=gpu_context,
+            model_name=model_name,
+            world_size=world_size,
+        )
+
+        layout_desc = get_layout_desc(gpu_context, self._ctx.chunk_size)
+        self._ctx.layout_desc_registry.register(model_name, world_size, layout_desc)
+
+        logger.info(
+            "Registered KV cache for GPU ID %d with %d layers",
+            instance_id,
+            gpu_context.num_layers,
+        )
+
+    def unregister_kv_cache(self, instance_id: int) -> None:
+        """Unregister the KV cache tensors for a given GPU instance ID.
+
+        Args:
+            instance_id: The GPU instance ID (such as PID).
+        """
+        entry = self._gpu_contexts.pop(instance_id, None)
+        if entry is None:
+            logger.warning(
+                "No registered GPU context found for instance ID %d", instance_id
+            )
+            return
+
+        self._ctx.layout_desc_registry.unregister(entry.model_name, entry.world_size)
+        logger.info("Unregistered KV cache for GPU ID %d", instance_id)
+        torch_dev.empty_cache()
+
+    @_lmcache_nvtx_annotate
+    def store(
+        self,
+        key: IPCCacheEngineKey,
+        instance_id: int,
+        gpu_block_ids: list[int],
+        event_ipc_handle: bytes,
+    ) -> tuple[bytes, bool]:
+        """Store the GPU KV cache blocks to CPU.
+
+        Args:
+            key: The IPC key for the KV cache blocks.
+                Must have worker_id != None (worker store operation).
+            instance_id: The GPU instance ID (such as PID).
+            gpu_block_ids: The GPU block IDs to store.
+            event_ipc_handle: The IPC handle of the event to wait on.
+
+        Returns:
+            A tuple where the first element is the IPC handle of the event
+            that signals the completion of the store operation, and the second
+            element indicates whether the store operation was successful.
+
+        Raises:
+            ValueError: If no GPU context is registered for the given instance ID.
+            RuntimeError: If the backend does not support IPC event handles.
+        """
+        st = time.perf_counter()
+        obj_keys = self._ctx.resolve_obj_keys(key)
+
+        entry = self._gpu_contexts.get(instance_id)
+        if entry is None:
+            raise ValueError(f"No GPU context registered for instance ID {instance_id}")
+        gpu_context = entry.gpu_context
+        model_name = entry.model_name
+
+        # ``blocks_per_chunk`` is counted in inference-engine-side
+        # blocks (each block addresses
+        # ``inference_engine_logical_block_size`` *logical* tokens).
+        # For compressed groups the per-group physical slot count
+        # differs, but the block-id indexing is shared with the engine
+        # and therefore uses the engine logical block size here.
+        blocks_per_chunk = (
+            self._ctx.chunk_size
+            // gpu_context.kv_layer_groups_manager.inference_engine_logical_block_size
+        )
+
+        with (
+            torch_dev.device(gpu_context.device),
+            torch_dev.stream(gpu_context.stream),
+        ):
+            check_interprocess_event_support()
+            event = torch_dev.Event(interprocess=True)
+
+            all_block_ids_gpu = gpu_context.stage_block_ids(gpu_block_ids)
+
+            if not hasattr(torch_dev.Event, "from_ipc_handle"):
+                raise RuntimeError(
+                    f"Backend '{torch_device_type}' does not support IPC event "
+                    "handles (Event.from_ipc_handle not available). "
+                    "Multiprocess IPC requires CUDA."
+                )
+            vllm_event = torch_dev.Event.from_ipc_handle(
+                gpu_context.device, event_ipc_handle
+            )
+            vllm_event.wait(stream=gpu_context.stream)
+
+            # CPU-synchronous sentinel: a GPU store is about to be enqueued.
+            # Must be published via publish() (not publish_on_stream) so the
+            # drain thread sees it before MP_REQUEST_END can race MP_STORE_END.
+            self._ctx.event_bus.publish(
+                Event(
+                    event_type=EventType.MP_STORE_SUBMITTED,
+                    session_id=key.request_id,
+                    metadata={"device": str(gpu_context.device)},
+                )
+            )
+
+            self._ctx.event_bus.publish_on_stream(
+                gpu_context.cupy_stream,
+                Event(
+                    event_type=EventType.MP_STORE_START,
+                    session_id=key.request_id,
+                    metadata={
+                        "device": str(gpu_context.device),
+                        "engine_id": instance_id,
+                        "model_name": model_name,
+                    },
+                ),
+            )
+
+            reserved_dict: dict[ObjectKey, MemoryObj] = {}
+            try:
+                layout_desc = get_layout_desc(gpu_context, self._ctx.chunk_size)
+                reserved_dict = self._ctx.storage_manager.reserve_write(
+                    obj_keys, layout_desc, "new"
+                )
+
+                # NOTE: Store is not batched because some obj_keys may be
+                # skipped (not in reserved_dict), making block_ids
+                # non-contiguous. Batching would require torch.cat to
+                # reassemble block_ids, negating the benefit.
+                num_groups = gpu_context.kv_layer_groups_manager.num_groups
+                for idx, obj_key in enumerate(obj_keys):
+                    if obj_key in reserved_dict:
+                        memory_obj = reserved_dict[obj_key]
+                    else:
+                        continue
+
+                    chunk_block_ids_gpu = all_block_ids_gpu[
+                        idx * blocks_per_chunk : (idx + 1) * blocks_per_chunk
+                    ]
+
+                    # Copy from GPU paged buffer to tmp buffer, then to CPU — per group
+                    for group_idx in range(num_groups):
+                        tmp_buffer = gpu_context.get_tmp_chunk_gpu_buffer(group_idx)
+                        group_kv_pointers = gpu_context.get_group_kv_pointers(group_idx)
+                        # Kernel contract: ``group_lmcache_chunk_size`` here is the
+                        # number of *physical* slots per chunk for this group
+                        # (= logical chunk_size // compress_ratio).
+                        group_lmcache_chunk_size = gpu_context.get_physical_chunk_size(
+                            group_idx
+                        )
+                        lmc_ops.multi_layer_block_kv_transfer(
+                            group_kv_pointers,
+                            [tmp_buffer.data_ptr()],
+                            chunk_block_ids_gpu,
+                            gpu_context.device,
+                            lmc_ops.TransferDirection.D2H,
+                            gpu_context.get_shape_desc(group_idx),
+                            group_lmcache_chunk_size,
+                            gpu_context.gpu_kv_format_,
+                            0,
+                        )
+                    # Store is not batched, so we always use chunk_idx=0 (single slot)
+                    lmcache_memcpy_async_d2h(
+                        gpu_context.get_tmp_gpu_buffer_flat(chunk_idx=0), memory_obj
+                    )
+            except Exception:
+                logger.exception("Cannot store keys due to exception")
+            finally:
+                event.record()
+                if reserved_dict:
+                    gpu_context.cupy_stream.launch_host_func(
+                        self._ctx.storage_manager.finish_write,
+                        list(reserved_dict.keys()),
+                    )
+                # All reserved MemoryObjs share one layout_desc, so per-object
+                # size is identical — avoid summing N identical values.
+                total_bytes = (
+                    next(iter(reserved_dict.values())).get_size() * len(reserved_dict)
+                    if reserved_dict
+                    else 0
+                )
+                self._ctx.event_bus.publish_on_stream(
+                    gpu_context.cupy_stream,
+                    Event(
+                        event_type=EventType.MP_STORE_END,
+                        session_id=key.request_id,
+                        metadata={
+                            "stored_count": len(reserved_dict),
+                            "device": str(gpu_context.device),
+                            "engine_id": instance_id,
+                            "model_name": model_name,
+                            "total_bytes": total_bytes,
+                        },
+                    ),
+                )
+
+        ed = time.perf_counter()
+        if length := len(reserved_dict):
+            logger.info(
+                "Stored %d tokens in %.3f seconds",
+                length * self._ctx.chunk_size,
+                ed - st,
+            )
+        return event.ipc_handle(), True
+
+    @_lmcache_nvtx_annotate
+    def retrieve(
+        self,
+        key: IPCCacheEngineKey,
+        instance_id: int,
+        gpu_block_ids: list[int],
+        event_ipc_handle: bytes,
+        skip_first_n_tokens: int = 0,
+    ) -> tuple[bytes, bool]:
+        """Retrieve the CPU KV cache and put into GPU blocks.
+
+        Args:
+            key: The IPC key for the KV cache blocks.
+                Must have worker_id != None (worker retrieve operation).
+            instance_id: The GPU instance ID (such as PID).
+            gpu_block_ids: The GPU block IDs to retrieve into.
+            event_ipc_handle: The IPC handle of the event to wait on.
+            skip_first_n_tokens: Number of tokens to skip writing at
+                the start of the retrieve range. This avoids overwriting
+                APC-shared GPU blocks that may be read concurrently by other
+                requests.
+
+        Returns:
+            A tuple where the first element is the IPC handle of the event
+            that signals the completion of the retrieve operation, and the
+            second element indicates whether the key was successfully retrieved.
+
+        Raises:
+            ValueError: If no GPU context is registered for the given instance ID.
+        """
+        st = time.perf_counter()
+        obj_keys = self._ctx.resolve_obj_keys(key)
+
+        entry = self._gpu_contexts.get(instance_id)
+        if entry is None:
+            raise ValueError(f"No GPU context registered for instance ID {instance_id}")
+        gpu_context = entry.gpu_context
+        model_name = entry.model_name
+
+        # CPU-synchronous sentinel: a GPU retrieve is about to be enqueued.
+        # Must be published via publish() (not publish_on_stream) so the
+        # drain thread sees it before MP_REQUEST_END can race MP_RETRIEVE_END.
+        self._ctx.event_bus.publish(
+            Event(
+                event_type=EventType.MP_RETRIEVE_SUBMITTED,
+                session_id=key.request_id,
+                metadata={"device": str(gpu_context.device)},
+            )
+        )
+
+        self._ctx.event_bus.publish_on_stream(
+            gpu_context.cupy_stream,
+            Event(
+                event_type=EventType.MP_RETRIEVE_START,
+                session_id=key.request_id,
+                metadata={
+                    "device": str(gpu_context.device),
+                    "engine_id": instance_id,
+                    "model_name": model_name,
+                },
+            ),
+        )
+
+        # ``skip_*_in_chunk`` is expressed in engine-block units
+        # (logical tokens), which is what the kernel's
+        # ``skip_blocks_in_chunk`` argument expects regardless
+        # of per-group compression.
+        ie_logical_block_size = (
+            gpu_context.kv_layer_groups_manager.inference_engine_logical_block_size
+        )
+        blocks_per_chunk = self._ctx.chunk_size // ie_logical_block_size
+
+        def _retrieve_loop(keys: list[ObjectKey], memory_objs: list[MemoryObj]) -> None:
+            _BATCH_SIZE = gpu_context.max_batch_size
+            num_groups = gpu_context.kv_layer_groups_manager.num_groups
+            for batch_idx, memory_obj_batch in enumerate(
+                batched_iteration(memory_objs, batch_size=_BATCH_SIZE)
+            ):
+                batch_len = len(memory_obj_batch)
+                chunk_start = batch_idx * self._ctx.chunk_size * _BATCH_SIZE
+                chunk_end = chunk_start + self._ctx.chunk_size * batch_len
+
+                effective_start = max(chunk_start, skip_first_n_tokens)
+                if effective_start >= chunk_end:
+                    # Entire batch is within APC range, skip it
+                    continue
+
+                skip_tokens_in_chunk = max(
+                    0,
+                    min(
+                        effective_start - chunk_start,
+                        self._ctx.chunk_size * batch_len - 1,
+                    ),
+                )
+                if skip_tokens_in_chunk % ie_logical_block_size != 0:
+                    logger.error(
+                        "skip_first_n_tokens (%d) is not aligned to "
+                        "inference_engine_logical_block_size (%d), "
+                        "rounding down from %d tokens to %d blocks",
+                        skip_first_n_tokens,
+                        ie_logical_block_size,
+                        skip_tokens_in_chunk,
+                        skip_tokens_in_chunk // ie_logical_block_size,
+                    )
+                skip_blocks_in_chunk = skip_tokens_in_chunk // ie_logical_block_size
+
+                start_chunk_id = batch_idx * _BATCH_SIZE
+                end_chunk_id = start_chunk_id + batch_len
+                chunk_block_ids_gpu = all_block_ids_gpu[
+                    start_chunk_id * blocks_per_chunk : end_chunk_id * blocks_per_chunk
+                ]
+
+                # Copy from CPU to GPU tmp buffers, then scatter to paged KV — per group
+                # H2D copy: each memory_obj maps to its own batch slot
+                for chunk_idx, memory_obj in enumerate(memory_obj_batch):
+                    lmcache_memcpy_async_h2d(
+                        memory_obj,
+                        gpu_context.get_tmp_gpu_buffer_flat(chunk_idx=chunk_idx),
+                    )
+                for group_idx in range(num_groups):
+                    tmp_buffers = gpu_context.get_tmp_chunk_gpu_buffer_batched(
+                        batch_len, group_idx
+                    )
+                    group_kv_pointers = gpu_context.get_group_kv_pointers(group_idx)
+                    group_lmcache_chunk_size = gpu_context.get_physical_chunk_size(
+                        group_idx
+                    )
+
+                    lmc_ops.multi_layer_block_kv_transfer(
+                        group_kv_pointers,
+                        [tb.data_ptr() for tb in tmp_buffers],
+                        chunk_block_ids_gpu,
+                        gpu_context.device,
+                        lmc_ops.TransferDirection.H2D,
+                        gpu_context.get_shape_desc(group_idx),
+                        group_lmcache_chunk_size,
+                        gpu_context.gpu_kv_format_,
+                        skip_blocks_in_chunk,
+                    )
+
+        with (
+            torch_dev.device(gpu_context.device),
+            torch_dev.stream(gpu_context.stream),
+        ):
+            # Stage all block_ids to GPU once before the loop
+            all_block_ids_gpu = gpu_context.stage_block_ids(gpu_block_ids)
+
+            check_interprocess_event_support()
+            event = torch_dev.Event(interprocess=True)
+
+            prefetched_keys: list[ObjectKey] = []
+            retrieve_succeeded = False
+            total_bytes = 0
+            try:
+                with self._ctx.storage_manager.read_prefetched_results(
+                    obj_keys
+                ) as memory_objs:
+                    if not memory_objs or len(memory_objs) != len(obj_keys):
+                        logger.error("Some keys not found during retrieve!")
+                        return event.ipc_handle(), False
+
+                    prefetched_keys = obj_keys[: len(memory_objs)]
+                    total_bytes = sum(mo.get_size() for mo in memory_objs)
+                    _retrieve_loop(obj_keys, memory_objs)
+                # Only set True when with-block exits normally
+                retrieve_succeeded = True
+            except Exception:
+                logger.exception("Cannot retrieve keys due to exception")
+                return event.ipc_handle(), False
+            finally:
+                event.record()
+                if retrieve_succeeded:
+                    gpu_context.cupy_stream.launch_host_func(
+                        self._ctx.storage_manager.finish_read_prefetched,
+                        prefetched_keys,
+                    )
+                self._ctx.event_bus.publish_on_stream(
+                    gpu_context.cupy_stream,
+                    Event(
+                        event_type=EventType.MP_RETRIEVE_END,
+                        session_id=key.request_id,
+                        metadata={
+                            "retrieved_count": len(prefetched_keys),
+                            "device": str(gpu_context.device),
+                            "engine_id": instance_id,
+                            "model_name": model_name,
+                            "cache_salt": key.cache_salt,
+                            "total_bytes": total_bytes,
+                        },
+                    ),
+                )
+        tokens_retrieved = len(obj_keys) * self._ctx.chunk_size
+        ed = time.perf_counter()
+        logger.info(
+            "Retrieved %d tokens in %.3f seconds",
+            tokens_retrieved,
+            ed - st,
+        )
+
+        return event.ipc_handle(), True
