+++ lmcache/v1/multiprocess/modules/lookup.py
@@ -1,0 +1,462 @@
+# SPDX-License-Identifier: Apache-2.0
+"""LookupModule: lookup, prefetch polling, and session lifecycle."""
+
+# Standard
+from dataclasses import dataclass
+from functools import partial
+import threading
+import time
+
+# First Party
+from lmcache.logging import init_logger
+from lmcache.v1.distributed.api import (
+    PrefetchHandle,
+    ipc_key_to_object_keys,
+)
+from lmcache.v1.mp_observability.event import Event, EventType
+from lmcache.v1.mp_observability.otel_init import register_gauge
+from lmcache.v1.multiprocess.custom_types import IPCCacheEngineKey
+from lmcache.v1.multiprocess.engine_context import MPCacheEngineContext
+from lmcache.v1.multiprocess.engine_module import (
+    HandlerSpec,
+    ThreadPoolType,
+)
+from lmcache.v1.multiprocess.protocol import RequestType
+from lmcache.v1.multiprocess.token_hasher import TokenHasher
+
+logger = init_logger(__name__)
+
+
+def compute_extra_count(
+    tp_size: int,
+    world_size: int,
+) -> int:
+    """Compute extra count for MLA multi-reader locking.
+
+    Non-MLA: each TP worker owns a distinct KV shard,
+      so each ObjectKey is retrieved by exactly 1
+      worker -> extra_count = 0.
+    MLA: TP does not split KV caches, all TP workers
+      share the same object. vLLM passes world_size
+      already divided by tp_size (e.g. world_size=1
+      for TP=4 PP=1), so ipc_keys_to_object_keys
+      only produces 1 ObjectKey per chunk.  All TP
+      workers retrieve that same ObjectKey, hence
+      extra_count = tp_size - 1.
+
+    Detection: tp > world_size means MLA (world_size
+    was divided by tp on the vLLM side).
+
+    Fallback: old vLLM (<= 0.8.5) does not send
+    tp_size (defaults to 1); we fall back to
+    world_size which gives extra_count = 0
+    (safe but may under-lock for MLA).
+
+    TODO: world_size currently carries an overloaded
+    meaning (total ranks for non-MLA vs total/tp for
+    MLA). Consider a dedicated field in the future.
+
+    Args:
+        tp_size: Tensor-parallel size from the client.
+        world_size: World size from the cache key.
+
+    Returns:
+        Number of extra count (0 for non-MLA).
+    """
+    tp = tp_size if tp_size > 1 else world_size
+    return tp - 1 if tp > world_size else 0
+
+
+@dataclass
+class _PrefetchJob:
+    handle: PrefetchHandle
+    world_size: int
+    request_id: str
+    # Number of tokens submitted for lookup (denominator for the L1+L2
+    # token-level hit-rate metric).  Equals ``len(chunk_hashes) * chunk_size``
+    # on the happy path; 0 for early-exit paths (no GPU context matches
+    # or chunk_hashes is empty).  Consumed at ``MP_LOOKUP_PREFETCH_END``
+    # emission time in ``query_prefetch_status``.
+    requested_tokens: int
+    # Captured at lookup time so the ``MP_LOOKUP_PREFETCH_END`` event can
+    # carry them as labels.  ``model_name`` lets dashboards slice hit rate
+    # per model in multi-model deployments; ``cache_salt`` slices per
+    # tenant / isolation domain (an empty string means no salt set).
+    model_name: str = ""
+    cache_salt: str = ""
+
+
+class LookupModule:
+    """Handles lookup, prefetch polling, lock release, and session lifecycle.
+
+    Owns the prefetch-job bookkeeping (``_prefetch_jobs``) and exposes
+    handlers for the LOOKUP, QUERY_PREFETCH_STATUS,
+    QUERY_PREFETCH_LOOKUP_HITS, FREE_LOOKUP_LOCKS, and END_SESSION
+    request types.
+
+    Args:
+        ctx: Shared engine context providing storage manager, token hasher,
+            session manager, event bus, layout descriptor registry, and
+            chunk size.
+    """
+
+    def __init__(self, ctx: MPCacheEngineContext) -> None:
+        self._ctx = ctx
+        self._prefetch_jobs: dict[str, _PrefetchJob] = {}
+        self._prefetch_job_lock = threading.Lock()
+        self._setup_metrics()
+
+    def get_handlers(self) -> list[HandlerSpec]:
+        """Return handler specs for all request types this module serves.
+
+        Returns:
+            List of handler specs for lookup-related request types.
+        """
+        return [
+            HandlerSpec(RequestType.LOOKUP, self.lookup, ThreadPoolType.NORMAL),
+            HandlerSpec(
+                RequestType.QUERY_PREFETCH_STATUS,
+                self.query_prefetch_status,
+                ThreadPoolType.NORMAL,
+            ),
+            HandlerSpec(
+                RequestType.QUERY_PREFETCH_LOOKUP_HITS,
+                self.query_prefetch_lookup_hits,
+                ThreadPoolType.NORMAL,
+            ),
+            HandlerSpec(
+                RequestType.FREE_LOOKUP_LOCKS,
+                self.free_lookup_locks,
+                ThreadPoolType.NORMAL,
+            ),
+            HandlerSpec(
+                RequestType.END_SESSION,
+                self.end_session,
+                ThreadPoolType.NORMAL,
+            ),
+        ]
+
+    def report_status(self) -> dict[str, int]:
+        """Return module-specific status information.
+
+        Returns:
+            Dictionary with the count of active prefetch jobs.
+        """
+        return {
+            "active_prefetch_jobs": self._active_prefetch_count(),
+        }
+
+    def close(self) -> None:
+        """Release resources owned by this module (no-op)."""
+        pass
+
+    # -----------------------------------------------------------------
+    # Handlers
+    # -----------------------------------------------------------------
+
+    def lookup(
+        self,
+        key: IPCCacheEngineKey,
+        tp_size: int,
+    ) -> None:
+        """Submit a prefix lookup.
+
+        Hashes the key, submits a prefetch task to the storage manager,
+        and registers the job under ``key.request_id`` for later polling
+        via query_prefetch_status.
+
+        Args:
+            key: Cache key with request_id embedded.
+            tp_size: Tensor-parallel size for MLA multi-reader locking.
+        """
+        model_name, world_size = key.model_name, key.world_size
+        self._ctx.event_bus.publish(
+            Event(
+                event_type=EventType.MP_REQUEST_START,
+                session_id=key.request_id,
+            )
+        )
+        self._ctx.event_bus.publish(
+            Event(
+                event_type=EventType.MP_LOOKUP_PREFETCH_START,
+                session_id=key.request_id,
+            )
+        )
+
+        layout_desc = self._ctx.layout_desc_registry.find(model_name, world_size)
+        if layout_desc is None:
+            logger.error(
+                "No GPU context found for model %s with world size %d during lookup!",
+                model_name,
+                world_size,
+            )
+            self._register_prefetch_job(
+                _PrefetchJob(
+                    handle=PrefetchHandle(
+                        prefetch_request_id=-1,
+                        external_request_id=key.request_id,
+                        l1_prefix_hit_count=0,
+                        total_requested_keys=0,
+                        submit_time=time.monotonic(),
+                    ),
+                    world_size=1,
+                    request_id=key.request_id,
+                    requested_tokens=0,
+                    model_name=model_name,
+                    cache_salt=key.cache_salt,
+                )
+            )
+            return
+
+        extra_count = compute_extra_count(tp_size, world_size)
+
+        chunk_hashes = self._ctx.token_hasher.compute_chunk_hashes(list(key.token_ids))
+        if not chunk_hashes:
+            self._register_prefetch_job(
+                _PrefetchJob(
+                    handle=PrefetchHandle(
+                        prefetch_request_id=-1,
+                        external_request_id=key.request_id,
+                        l1_prefix_hit_count=0,
+                        total_requested_keys=0,
+                        submit_time=time.monotonic(),
+                    ),
+                    world_size=1,
+                    request_id=key.request_id,
+                    requested_tokens=0,
+                    model_name=model_name,
+                    cache_salt=key.cache_salt,
+                )
+            )
+            return
+
+        # Total chunk-aligned tokens submitted for lookup; surfaces as the
+        # denominator of the L1+L2 token-level hit-rate via the
+        # ``requested_tokens`` field on ``MP_LOOKUP_PREFETCH_END``.  Sub-chunk
+        # trailing tokens are intentionally excluded — they cannot hit at
+        # chunk granularity.
+        requested_tokens = len(chunk_hashes) * self._ctx.chunk_size
+
+        # Guard with has_subscribers() to avoid allocating the metadata dict
+        # (including dtype/shape list comprehensions) when no subscriber is
+        # listening (e.g. lookup hash logger is disabled).
+        if self._ctx.event_bus.has_subscribers(EventType.MP_LOOKUP):
+            self._ctx.event_bus.publish(
+                Event(
+                    event_type=EventType.MP_LOOKUP,
+                    session_id=key.request_id,
+                    metadata={
+                        "request_id": key.request_id,
+                        "chunk_hashes": chunk_hashes,
+                        "model_name": model_name,
+                        "chunk_size": self._ctx.chunk_size,
+                        "seq_len": len(key.token_ids),
+                        "dtypes": [str(d) for d in layout_desc.dtypes],
+                        "shapes": [list(s) for s in layout_desc.shapes],
+                    },
+                )
+            )
+
+        session = self._ctx.session_manager.get_or_create(key.request_id)
+        session.set_tokens(list(key.token_ids))
+        session.lookup_ipc_key = key
+
+        obj_keys = ipc_key_to_object_keys(key, chunk_hashes)
+
+        handle = self._ctx.storage_manager.submit_prefetch_task(
+            obj_keys,
+            layout_desc,
+            extra_count=extra_count,
+            external_request_id=key.request_id,
+        )
+        self._register_prefetch_job(
+            _PrefetchJob(
+                handle=handle,
+                world_size=key.world_size,
+                request_id=key.request_id,
+                requested_tokens=requested_tokens,
+                model_name=model_name,
+                cache_salt=key.cache_salt,
+            )
+        )
+
+    def query_prefetch_lookup_hits(
+        self,
+        request_id: str,
+    ) -> int | None:
+        """Query the number of hits for a prefetch request before it's finished.
+
+        Args:
+            request_id: The external request ID passed in the lookup key.
+
+        Returns:
+            The number of hits for the prefetched keys if the lookup phase is
+            done. None if the lookup phase is still in progress. 0 if the
+            request_id is unknown (already completed and consumed, or invalid).
+        """
+        with self._prefetch_job_lock:
+            job = self._prefetch_jobs.get(request_id)
+
+        if job is None:
+            logger.warning(
+                "Prefetch job for request %s not found (already completed or invalid)",
+                request_id,
+            )
+            return 0
+
+        found_count = self._ctx.storage_manager.query_prefetch_lookup_hits(job.handle)
+        if found_count is None:
+            return None
+
+        found_count = found_count // job.world_size
+        return found_count
+
+    def query_prefetch_status(
+        self,
+        request_id: str,
+    ) -> int | None:
+        """Poll the status of a prefetch job by request_id.
+
+        Returns the chunk count when the prefetch is complete, or None
+        if it is still in progress.  The job entry is automatically
+        removed once a non-None result is returned (exactly-once
+        semantics).
+
+        Args:
+            request_id: The external request ID passed in the lookup key.
+
+        Returns:
+            Chunk count (int) when done, None if still in progress,
+            0 if the request_id is unknown (already completed and consumed,
+            or invalid).
+        """
+        with self._prefetch_job_lock:
+            job = self._prefetch_jobs.get(request_id)
+        if job is None:
+            logger.warning(
+                "Prefetch job for request %s not found (already completed or invalid)",
+                request_id,
+            )
+            return 0
+
+        found_count = self._ctx.storage_manager.query_prefetch_status(job.handle)
+        if found_count is None:
+            return None
+
+        # NOTE(Kuntai): this assumes two things:
+        # 1. the world size is the same between keys
+        # 2. the lookup sort the keys in prefix order and breaks at the
+        #    first failure
+        found_count = found_count // job.world_size
+
+        self._ctx.event_bus.publish(
+            Event(
+                event_type=EventType.MP_LOOKUP_PREFETCH_END,
+                session_id=job.request_id,
+                metadata={
+                    "found_count": found_count,
+                    "requested_tokens": job.requested_tokens,
+                    "hit_tokens": found_count * self._ctx.chunk_size,
+                    "model_name": job.model_name,
+                    "cache_salt": job.cache_salt,
+                },
+            )
+        )
+
+        with self._prefetch_job_lock:
+            self._prefetch_jobs.pop(request_id, None)
+
+        return found_count
+
+    def free_lookup_locks(
+        self,
+        key: IPCCacheEngineKey,
+        tp_size: int,
+    ) -> None:
+        """Release read locks acquired during lookup.
+
+        Hashes are computed only for chunks in ``[start, end)`` to avoid
+        unnecessary work on tokens outside that range.
+        ``start`` and ``end`` must be aligned to ``chunk_size``; it is the
+        caller's responsibility to align the boundaries as desired.
+
+        Computes the extra reader count from ``tp_size`` and
+        ``world_size`` the same way :meth:`lookup` does, so
+        the correct number of locks is released.
+
+        Args:
+            key: Cache key whose read locks should be released.
+            tp_size: Tensor-parallel size for MLA
+                multi-reader locking.
+        """
+        chunk_hashes = self._ctx.token_hasher.compute_chunk_hashes(
+            list(key.token_ids), start=key.start, end=key.end
+        )
+        if not chunk_hashes:
+            return
+        obj_keys = ipc_key_to_object_keys(key, chunk_hashes)
+
+        extra_count = compute_extra_count(tp_size, key.world_size)
+
+        self._ctx.storage_manager.finish_read_prefetched(
+            obj_keys, extra_count=extra_count
+        )
+
+    def end_session(self, request_id: str) -> None:
+        """Remove the session for a finished request.
+
+        Args:
+            request_id: The request ID whose session should be removed.
+        """
+        self._ctx.event_bus.publish(
+            Event(
+                event_type=EventType.MP_VLLM_END_SESSION,
+                metadata={"request_id": request_id},
+            )
+        )
+        session = self._ctx.session_manager.remove(request_id)
+        self._ctx.event_bus.publish(
+            Event(
+                event_type=EventType.MP_REQUEST_END,
+                session_id=request_id,
+            )
+        )
+        if session is None:
+            logger.warning("Session %s not found, skipping touch", request_id)
+            return
+        if session.lookup_ipc_key is None:
+            logger.warning(
+                "Session %s has no lookup ipc key, skipping touch",
+                request_id,
+            )
+            return
+
+        chunk_hashes = [TokenHasher.hash_to_bytes(h) for h in session.get_hashes(0)]
+        obj_keys = ipc_key_to_object_keys(session.lookup_ipc_key, chunk_hashes)
+        # unified touch of all keys, which include retrieved and stored keys
+        # TODO(chunxiaozheng): when l2 is enabled, the prefetched keys from l2 are temp
+        #  and will be deleted after finish_read_prefetched, when we touch all keys,
+        #  these keys has been deleted and will not be touched.
+        self._ctx.storage_manager.touch_l1_keys(obj_keys)
+
+    # -----------------------------------------------------------------
+    # Internal helpers
+    # -----------------------------------------------------------------
+
+    def _register_prefetch_job(self, job: _PrefetchJob) -> None:
+        with self._prefetch_job_lock:
+            self._prefetch_jobs[job.request_id] = job
+
+    def _active_prefetch_count(self) -> int:
+        """Return the number of active prefetch jobs (thread-safe)."""
+        with self._prefetch_job_lock:
+            return len(self._prefetch_jobs)
+
+    def _setup_metrics(self) -> None:
+        """Register OTel observable gauges for lookup module metrics."""
+        _gauge = partial(register_gauge, "lmcache.mp_engine")
+        _gauge(
+            "lmcache_mp.active_prefetch_jobs",
+            "Number of active prefetch jobs",
+            self._active_prefetch_count,
+        )
